feat: split create data assests unit tests into two new ones and reworked the database to allow duplicates in tables
