+++ scripts/database/create_tables.sql
@@ -43,9 +43,8 @@ IF NOT EXISTS
 BEGIN
     CREATE TABLE PARTICIPANT_SCREENING_PROFILE
     (
-      NHS_NUMBER NVARCHAR (50) NOT NULL,
-        CONSTRAINT PK_NHS_NUMBER
-          PRIMARY KEY (NHS_NUMBER),
+      ID INT IDENTITY(1,1) PRIMARY KEY,
+      NHS_NUMBER                             NVARCHAR (50) NOT NULL,
       SCREENING_NAME                         VARCHAR(50) NULL,
       PRIMARY_CARE_PROVIDER                  VARCHAR(50) NULL,
       PREFERRED_LANGUAGE                     VARCHAR(50) NULL,
@@ -77,9 +76,8 @@ IF NOT EXISTS
 BEGIN
     CREATE TABLE PARTICIPANT_SCREENING_EPISODE
     (
-      EPISODE_ID NVARCHAR (50) NOT NULL,
-        CONSTRAINT PK_EPISODE_ID
-          PRIMARY KEY (EPISODE_ID),
+      ID INT IDENTITY(1,1) PRIMARY KEY,
+      EPISODE_ID                       NVARCHAR (50) NOT NULL,
       SCREENING_NAME                   VARCHAR(50) NULL,
       NHS_NUMBER                       VARCHAR(50) NULL,
       EPISODE_TYPE                     VARCHAR(50) NULL,
