+++ tests/BIAnalyticsServiceTests/CreateParticipantScreeningEpisodeTests/CreateParticipantScreeningEpisodeTests.csproj
@@ -23,7 +23,7 @@
 
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\src\BIAnalyticsService\CreateDataAssets\CreateDataAssets.csproj" />
+    <ProjectReference Include="..\..\..\src\BIAnalyticsService\CreateParticipantScreeningEpisode\CreateParticipantScreeningEpisode.csproj" />
     <ProjectReference Include="..\..\TestUtils\TestUtils.csproj" />
   </ItemGroup>
